[FEAT] DetailVC가 HomeVC에서 선택한 cell의 데이터와 연결되도록 수정(#6)
